Extract generic wrapper
